# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Portfolio API",
    "description": "Backend API for portfolio showcase with user management, projects, skills, and analytics",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "hyoukjoo@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Get all users",
        "description": "Retrieve a list of all users in the system",
        "responses": {
          "200": {
            "description": "List of users",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                },
                "count": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create a new user",
        "description": "Create a new user with the provided information",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by their ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update user",
        "description": "Update an existing user by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete user",
        "description": "Delete a user by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "이혁주"
        },
        "email": {
          "type": "string",
          "example": "hyoukjoo@example.com"
        },
        "role": {
          "type": "string",
          "example": "Backend Developer"
        },
        "avatar": {
          "type": "string",
          "example": "https://avatars.githubusercontent.com/u/12345678"
        },
        "bio": {
          "type": "string",
          "example": "Backend engineer specializing in Go and cloud architecture"
        },
        "website": {
          "type": "string",
          "example": "https://hyoukjoolee.github.io/portfolio"
        },
        "location": {
          "type": "string",
          "example": "Seoul, Korea"
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["Go", "JavaScript", "Docker", "AWS"]
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CreateUserRequest": {
      "type": "object",
      "required": ["name", "email", "role"],
      "properties": {
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "role": {
          "type": "string",
          "example": "developer"
        },
        "avatar": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "UpdateUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}